reflection support, split up interpreter
