+++ packages/core/src/ast/assignmentExpression.ts
@@ -1,7 +1,7 @@
 import { CompletionExpressionMetadata } from "./expressionMetadata.js";
 import { Expression } from "./expression.js";
 import { WrapperObject } from "../runtime/objects/wrapperObject.js";
-import { InterpreterContext } from "../runtime/interpreter.js";
+import { InterpreterContext } from "../runtime/interpreter/interpreterContext.js";
 
 /**
  * Assignment Expression
