+++ packages/core/src/ast/invocationExpression.ts
@@ -2,7 +2,7 @@ import { ExpressionMetadata } from "./expressionMetadata.js";
 import { Expression } from "./expression.js";
 import { AbstractInvocationExpression } from "./abstractInvocationExpression.js";
 import { ListEntry } from "./listEntry.js";
-import { InterpreterContext } from "../runtime/interpreter.js";
+import { InterpreterContext } from "../runtime/interpreter/interpreterContext.js";
 import { WrapperObject } from "../runtime/objects/wrapperObject.js";
 
 /*
