+++ packages/core/src/ast/listEntry.ts
@@ -1,4 +1,4 @@
-import { InterpreterContext } from "../runtime/interpreter.js";
+import { InterpreterContext } from "../runtime/interpreter/interpreterContext.js";
 import { WrapperObject, WrapperObjectFieldRetriever } from "../runtime/objects/wrapperObject.js";
 import { Expression } from "./expression.js";
