+++ packages/core/src/ast/selfInvocationExpression.ts
@@ -2,7 +2,7 @@ import { CompletionExpressionMetadata } from "./expressionMetadata.js";
 import { Expression } from "./expression.js";
 import { AbstractInvocationExpression } from "./abstractInvocationExpression.js";
 import { ListEntry } from "./listEntry.js";
-import { InterpreterContext } from "../runtime/interpreter.js";
+import { InterpreterContext } from "../runtime/interpreter/interpreterContext.js";
 import { WrapperObject } from "../runtime/objects/wrapperObject.js";
 
 /**
