+++ packages/core/src/index.ts
@@ -46,8 +46,11 @@ export * from "./runtime/objects/literalObject.js";
 export * from "./runtime/objects/nullObject.js";
 export * from "./runtime/objects/numberObject.js";
 export * from "./runtime/objects/stringObject.js";
+export * from "./runtime/objects/wrapperObject.js";
 export * from "./runtime/executableAstHelper.js";
-export * from "./runtime/interpreter.js";
+export * from "./runtime/interpreter/interpreter.js";
+export * from "./runtime/interpreter/interpreterContext.js";
+export * from "./runtime/interpreter/interpreterModule.js";
 export * from "./runtime/runtimeAstTransformer.js";
 export * from "./runtime/runtimeError.js";
 export * from "./runtime/semanticFieldNames.js";
