+++ packages/core/src/runtime/ast/executableAbstractFunctionExpression.ts
@@ -1,6 +1,6 @@
 import { Expression } from "../../ast/expression.js";
 import { Type } from "../../types/base.js";
-import { InterpreterContext } from "../interpreter.js";
+import { InterpreterContext } from "../interpreter/interpreterContext.js";
 import { BaseObject } from "../objects/baseObject.js";
 import { ExecutableExpression } from "./executableExpression.js";
