+++ packages/core/src/runtime/ast/executableBracketExpression.ts
@@ -1,5 +1,5 @@
 import { BracketExpression } from "../../ast/bracketExpression.js";
-import { InterpreterContext } from "../interpreter.js";
+import { InterpreterContext } from "../interpreter/interpreterContext.js";
 import { FieldEntry } from "../objects/baseObject.js";
 import { ExecutableExpression } from "./executableExpression.js";
