+++ packages/core/src/runtime/ast/executableConstExpression.ts
@@ -1,4 +1,4 @@
-import { InterpreterContext } from "../interpreter.js";
+import { InterpreterContext } from "../interpreter/interpreterContext.js";
 import { FieldEntry } from "../objects/baseObject.js";
 import { ExecutableExpression } from "./executableExpression.js";
