+++ packages/core/src/runtime/ast/executableFieldAccessExpression.ts
@@ -1,5 +1,5 @@
 import { FieldAccessExpression } from "../../ast/fieldAccessExpression.js";
-import { InterpreterContext } from "../interpreter.js";
+import { InterpreterContext } from "../interpreter/interpreterContext.js";
 import { FieldEntry } from "../objects/baseObject.js";
 import { ExecutableExpression } from "./executableExpression.js";
