+++ packages/core/src/runtime/ast/executableStringLiteralExpression.ts
@@ -1,5 +1,5 @@
 import { StringLiteralExpression } from "../../ast/stringLiteralExpression.js";
-import { InterpreterContext } from "../interpreter.js";
+import { InterpreterContext } from "../interpreter/interpreterContext.js";
 import { FieldEntry } from "../objects/baseObject.js";
 import { StringObject } from "../objects/stringObject.js";
 import { ExecutableLiteralExpression } from "./executableLiteralExpression.js";
