+++ packages/core/src/runtime/executableAstHelper.ts
@@ -6,7 +6,7 @@ import { ExecutableIdentifierExpression } from "./ast/executableIdentifierExpres
 import { ExecutableNativeFunctionExpression, NativeFunctionType } from "./ast/executableNativeFunctionExpression.js";
 import { ExecutableNumberLiteralExpression } from "./ast/executableNumberLiteralExpression.js";
 import { ExecutableStringLiteralExpression } from "./ast/executableStringLiteralExpression.js";
-import { InterpreterContext } from "./interpreter.js";
+import { InterpreterContext } from "./interpreter/interpreterContext.js";
 import { BaseObject, FieldEntry } from "./objects/baseObject.js";
 import { FullObject } from "./objects/fullObject.js";
 import { generateArgs } from "./objects/functionObject.js";
