+++ packages/core/src/stdlib/modules/commonModule.ts
@@ -1,7 +1,7 @@
 import { FunctionExpression } from "../../ast/functionExpression.js";
 import { ExecutableConstExpression } from "../../runtime/ast/executableConstExpression.js";
 import { assign, jsFun, native } from "../../runtime/executableAstHelper.js";
-import { InterpreterModule } from "../../runtime/interpreter.js";
+import { InterpreterModule } from "../../runtime/interpreter/interpreterModule.js";
 import { FieldEntry } from "../../runtime/objects/baseObject.js";
 import { RuntimeError } from "../../runtime/runtimeError.js";
 import { SemanticFieldNames } from "../../runtime/semanticFieldNames.js";
