+++ packages/core/src/stdlib/modules/stringModule.ts
@@ -1,6 +1,6 @@
 import { ExecutableNativeExpression } from "../../runtime/ast/executableNativeExpression.js";
 import { assign, fun, id, jsFun } from "../../runtime/executableAstHelper.js";
-import { InterpreterModule } from "../../runtime/interpreter.js";
+import { InterpreterModule } from "../../runtime/interpreter/interpreterModule.js";
 import { StringObject } from "../../runtime/objects/stringObject.js";
 import { SemanticFieldNames } from "../../runtime/semanticFieldNames.js";
 import { stringType } from "../../types/string.js";
